Move local/global env code from exe.c to env.c
